@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+	selector: 'app-root',
+	template: `
+		<router-outlet></router-outlet>
+	`,
+	styleUrls: ['./app.component.css']
 })
-export class AppComponent {
-  title = 'kat-fact';
-}
+export class AppComponent { }
